LPS-56302 Move WikiItemSelector Criterion and Handler to wiki API
